Fix README example
This commit fixes a typo in the README: `create-db` _requires_ two
arguments (a database and an options map), but the README example
includes only the former argument.  This causes an `ArityException`:

```
ArityException Wrong number of args (1) passed to: clj-leveldb$create-db
```

This commit fixes the README example to have the right number of args.